Allow sysadmins to see what is being collected
This allows system admins to see what is being gathered on their
endpoints. It can be enabled by creating a directory named `uploads`
in the same directory as the `local_buffer` file. If this directory
exists then VQL responses that are uploaded to the server will also
be appended to a daily JSON file within this directory. There is no
auto removal of these files so admins will have to manage that
themselves. It can be disabled by simply removing the directory.